pulled a txt file, not a bin; fixed

+++ bash.sh
@@ -4,18 +4,26 @@ source packages.conf
 
 echo "Works fine"
 
+NVIM_V="v0.10.0"
+
 sudo -v
 # checking if lazyvim is installed
 if dpkg -s neovim >/dev/null 2>&1; then
     echo "nvim is configurated"
 else
     echo -e "nvim is not configurated\n     Installing..."
-    curl -LO https://github.com/neovim/neovim/releases/download/v0.11.1/nvim.appimage
+    curl -fLo nvim.appimage "https://github.com/neovim/neovim/releases/download/${NVIM_V}/nvim.appimage"
+
+    if [[ ! -f nvim.appimage ]]; then
+        echo "Error in pulling ${NVIM_V} neovim. Exiting..."
+        exit 1
+    fi
+
     chmod u+x nvim.appimage
     sudo mv nvim.appimage /usr/local/bin/nvim
 fi
 
-if [[ -d "${HOME}/.config" ]]; then
+if [[ ! -d "${HOME}/.config" ]]; then
     echo "Making the .config file"
     mkdir "${HOME}/.config"
 fi
